Consolidate familiar studio to five tabs: merge Look + lifecycle into Identity
Settings → Familiars had eight studio tabs; three carried overlapping or
relocated concerns. The strip is now Identity | Brain | Memory | Projects |
Vault:

- Identity absorbs the Look sections (avatar, icon, backdrop, accent) by
  hosting FamiliarStudioLookTab inline, followed by a new lifecycle section.
- New familiar-lifecycle-section.tsx keeps ONLY archive/unarchive and the
  undo-safe remove flow (confirm copy, UndoToast, Recently removed shelf,
  tombstone-backed restore) for the selected familiar. Roster reorder,
  reset-all-overrides, and the Open Appearance link retired with the tab.
- Journal studio tab removed — the journal lives in Grimoire (Memories →
  Journal); journal-entries.tsx drops its now-orphaned `standalone` fork.
- Context: FamiliarStudioTab union/STUDIO_TABS shrink; listView state
  removed; openFamiliarStudioListView now just opens Settings → Familiars
  (stale persisted tab values fall back to Identity via the existing guard).
- Deep links retargeted: Familiars-view "Remove familiar…" → identity tab;
  settings search rows for Look/Lifecycle/Journal folded into Identity.
- Tests renamed/retargeted to pin the new intent (journal-redirect.test.ts,
  familiar-lifecycle-section.test.ts); route-seam pins kept verbatim.

Verified: app suite 891 files green, typecheck, pnpm lint, design-token
ratchets, check-tests-wired.
