+++ scripts/run-tests.mjs
@@ -644,9 +644,9 @@ export const SUITES = {
     "src/components/familiar-glyph-picker-panel.test.ts",
     "src/components/familiar-glyph-loading.test.ts",
     "src/components/familiar-studio-brain-tab.test.ts",
-    "src/components/familiar-studio-journal-tab.test.ts",
+    "src/components/journal-redirect.test.ts",
     "src/components/familiar-studio-identity-tab.test.ts",
-    "src/components/familiar-studio-lifecycle-tab.test.ts",
+    "src/components/familiar-lifecycle-section.test.ts",
     "src/components/github-task-field.test.ts",
     "src/components/asana-task-field.test.ts",
     "src/lib/task-asana.test.ts",
