+++ src/components/journal-redirect.test.ts
@@ -1,7 +1,9 @@
 // @ts-nocheck
-// Journal lives in the Familiar Studio (Settings → Familiars → Journal).
-// Source-scan invariants for the tab wiring and the redirect from the old
-// top-level Journal surface.
+// The Journal lives in the Grimoire (Memories → Journal tab). Both prior homes
+// — the top-level Journal page and the Settings → Familiars studio tab — are
+// retired. These pins hold the redirect seams (mode remap, palette
+// reachability, restore path) and the invariant that neither retired surface
+// creeps back.
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 
@@ -9,17 +11,12 @@ const read = (rel) => readFileSync(new URL(rel, import.meta.url), "utf8");
 
 const ctx = read("../lib/familiar-studio-context.tsx");
 
-// ── Studio context knows the journal tab ─────────────────────────────────────
-assert.match(ctx, /"journal"/, "FamiliarStudioTab union includes journal");
-assert.match(
-  ctx,
-  /STUDIO_TABS: readonly FamiliarStudioTab\[\][\s\S]*?"journal"/,
-  "the canonical tab list includes journal",
-);
+// ── Studio context: journal is no longer a studio tab ────────────────────────
+assert.doesNotMatch(ctx, /"journal"/, "FamiliarStudioTab union no longer includes journal");
 assert.match(
   ctx,
   /\(STUDIO_TABS as readonly string\[\]\)\.includes\(stored \?\? ""\)/,
-  "the persisted-tab restore guard checks against STUDIO_TABS",
+  "the persisted-tab restore guard checks against STUDIO_TABS (a stale stored \"journal\" falls back to the default tab)",
 );
 // One shared redirect helper: workspace surfaces and the redirecting provider
 // both route through it, so the tab/familiar handoff keys can't drift.
@@ -34,42 +31,13 @@ assert.match(
   "the redirecting provider reuses the helper",
 );
 
-const wrapper = read("./familiar-studio-journal-tab.tsx");
 const inline = read("./familiar-studio-inline.tsx");
 const sections = read("./settings-sections.ts");
-const css = read("../styles/journal.css");
 
-// ── Wrapper: reuse JournalEntries pinned to the studio's familiar ────────────
-assert.match(wrapper, /import "@\/styles\/journal\.css"/, "wrapper carries the journal styles");
-assert.match(wrapper, /<JournalEntries/, "wrapper renders the existing JournalEntries surface");
-assert.match(
-  wrapper,
-  /useMemo\(\(\) => new Set\(\[familiar\.id\]\), \[familiar\.id\]\)/,
-  "the multiselect scope is pinned to the one familiar being edited",
-);
-assert.match(wrapper, /activeFamiliarId=\{familiar\.id\}/, "generation targets the studio familiar");
-
-// ── Inline panel: the tab is registered and rendered ─────────────────────────
-assert.match(
-  inline,
-  /\{ id: "journal", label: "Journal", icon: "ph:book-open" \}/,
-  "the studio tab bar includes Journal",
-);
-assert.match(
-  inline,
-  /activeTab === "journal" \? <FamiliarStudioJournalTab familiar=\{familiar\} allFamiliars=\{resolved\} \/> : null/,
-  "the journal tab body renders the wrapper",
-);
-
-// ── Settings search reaches the tab ──────────────────────────────────────────
-assert.match(sections, /familiarTab: "journal"/, "the journal studio tab is indexed for settings search");
-
-// ── Studio host gives the master-detail journal a bounded height ─────────────
-assert.match(
-  css,
-  /\.familiar-studio-journal \.journal-list \{[\s\S]*?height:/,
-  "journal-list gets an explicit height inside the studio body",
-);
+// ── Inline panel + settings search: no journal tab wiring remains ────────────
+assert.doesNotMatch(inline, /"journal"/, "the studio tab bar no longer includes Journal");
+assert.doesNotMatch(inline, /FamiliarStudioJournalTab/, "the journal tab wrapper is gone");
+assert.doesNotMatch(sections, /familiarTab: "journal"/, "settings search no longer indexes a journal studio tab");
 
 const entriesSrc = read("./journal/journal-entries.tsx");
 
@@ -85,14 +53,18 @@ assert.match(
   "hasEntry honors the familiar scope",
 );
 
+// The Settings host is gone, and its `standalone` fork went with it — the
+// Grimoire (workspace) is the only journal host, so the workspace event bus is
+// always available.
+assert.doesNotMatch(entriesSrc, /standalone/, "journal-entries no longer carries the settings-host fork");
+
 const ws = read("./workspace.tsx");
 const sidebar = read("./sidebar-minimal.tsx");
 const pageDrag = read("../lib/page-drag.ts");
 const slash = read("../lib/slash-commands.ts");
 
 // ── Workspace: "journal" is a redirect-only mode (like groupchat) ────────────
-// It now opens the Grimoire surface on its Journal tab; per-familiar journals
-// still live in Settings → Familiars → Journal via FamiliarStudioJournalTab.
+// It opens the Grimoire surface on its Journal tab.
 assert.match(
   ws,
   /if \(next === "journal"\) \{[\s\S]{0,400}?setGrimoireView\("journal"\);\s*\n\s*setModeRaw\("grimoire"\)/,
@@ -113,7 +85,7 @@ assert.doesNotMatch(sidebar, /generated sketches/, "sidebar description no longe
 assert.match(pageDrag, /NON_SPLITTABLE = new Set\(\["terminal", "journal"\]\)/, "journal is excluded from drag-to-split");
 
 // ── Slash palette copy matches the new home ───────────────────────────────────
-assert.match(slash, /name: "\/journal"[^}]*Settings/, "/journal description points at Settings");
+assert.match(slash, /name: "\/journal"/, "the /journal slash command survives the redirect");
 assert.doesNotMatch(slash, /Journal's Canvas tab/, "/canvas no longer advertises the Canvas page");
 
 const artifactViewer = read("./chat-artifact-viewer.tsx");
@@ -140,22 +112,9 @@ const ghReview = read("./gh-review-actions.tsx");
 assert.doesNotMatch(ghReview, /cave:canvas:layer|mode: "canvas"/, "PR review export no longer jumps to the retired Canvas page");
 assert.match(ghReview, /openArtifactHtml\(artifact\.code\)/, "exported review artifacts open directly in a browser tab");
 
-// ── Settings host: dead workspace events are avoided ─────────────────────────
-assert.match(wrapper, /standalone/, "the studio tab renders JournalEntries in standalone mode");
-assert.match(
-  entriesSrc,
-  /AUTOMATIONS\.filter\(\(a\) => a\.action !== "run"\)/,
-  "standalone mode hides the chat-handoff automation",
-);
-assert.match(
-  entriesSrc,
-  /window\.location\.assign\(`\/\?mode=\$\{encodeURIComponent\(notice\.action!\.mode\)\}`\)/,
-  "standalone toast actions deep-link back into the workspace",
-);
-
 // ── One-entry-per-day store: no silent cross-familiar overwrite ──────────────
 assert.match(entriesSrc, /const outOfScopeBy =/, "derives the out-of-scope author");
 assert.match(entriesSrc, /if \(outOfScopeBy\) return;/, "generate refuses to overwrite an out-of-scope entry");
 assert.match(entriesSrc, /written by \$\{outOfScopeBy\}/, "the empty state names the actual author instead of inviting an overwrite");
 
-console.log("familiar-studio-journal-tab.test.ts: ok");
+console.log("journal-redirect.test.ts: ok");
